new_llm_completion

new_llm_completion

Usage

new_llm_completion(
  x = character(),
  Call = NULL,
  Prompt = NULL,
  Model = NULL,
  Model_Provider = NULL,
  Date = NULL,
  Temperature = NULL,
  Raw = NULL
)

Arguments

x

The vector to be made into a new llm_completion object

Prompt

The prompt used to generate the completion

Model

The model used to generate the completion

Model_Provider

The provider of the model used to generate the completion

Date

The date the completion was generated

Temperature

The temperature used to generate the completion

Raw

A sample of the raw completion data

Value

A new llm_completion object